Trait equistore::Array

pub trait Array: Any + Send + Sync {
    fn as_any(&self) -> &dyn Any;
    fn as_any_mut(&mut self) -> &mut dyn Any;
    fn origin(&self) -> DataOrigin;
    fn create(&self, shape: &[usize]) -> Box<dyn Array>;
    fn copy(&self) -> Box<dyn Array>;
    fn data(&self) -> &[f64];
    fn shape(&self) -> &[usize];
    fn reshape(&mut self, shape: &[usize]);
    fn swap_axes(&mut self, axis_1: usize, axis_2: usize);
    fn move_samples_from(
        &mut self, 
        input: &dyn Array, 
        samples: &[eqs_sample_mapping_t], 
        properties: Range<usize>
    );
}

A rust trait with the same interface as eqs_array_t, see this struct for documentation on all methods.

Required Methods

fn as_any(&self) -> &dyn Any

fn as_any_mut(&mut self) -> &mut dyn Any

fn origin(&self) -> DataOrigin

fn create(&self, shape: &[usize]) -> Box<dyn Array>

fn copy(&self) -> Box<dyn Array>

fn data(&self) -> &[f64]

fn shape(&self) -> &[usize]

fn reshape(&mut self, shape: &[usize])

fn swap_axes(&mut self, axis_1: usize, axis_2: usize)

fn move_samples_from(
&mut self,
input: &dyn Array,
samples: &[eqs_sample_mapping_t],
properties: Range<usize>
)